pc-vision/requirements.txt
opencv-python
paho-mqtt
numpy